+++ packages/pds/src/actor-store/actor-store-resources.ts
@@ -1,0 +1,8 @@
+import { BlobStore } from '@atproto/repo'
+import { BackgroundQueue } from '../background'
+
+export type ActorStoreResources = {
+  blobstore: (did: string) => BlobStore
+  backgroundQueue: BackgroundQueue
+  reservedKeyDir?: string
+}
