+++ packages/pds/src/api/com/atproto/repo/putRecord.ts
@@ -1,21 +1,22 @@
-import { CID } from 'multiformats/cid'
+import { BlobRef } from '@atproto/lexicon'
+import { CommitData } from '@atproto/repo'
 import { AtUri } from '@atproto/syntax'
 import { AuthRequiredError, InvalidRequestError } from '@atproto/xrpc-server'
-import { CommitData } from '@atproto/repo'
-import { BlobRef } from '@atproto/lexicon'
-import { Server } from '../../../../lexicon'
-import { prepareUpdate, prepareCreate } from '../../../../repo'
+import { CID } from 'multiformats/cid'
+import { ActorStoreTransactor } from '../../../../actor-store/actor-store-transactor'
 import AppContext from '../../../../context'
+import { Server } from '../../../../lexicon'
+import { ids } from '../../../../lexicon/lexicons'
+import { Record as ProfileRecord } from '../../../../lexicon/types/app/bsky/actor/profile'
 import {
   BadCommitSwapError,
   BadRecordSwapError,
   InvalidRecordError,
+  prepareCreate,
   PreparedCreate,
   PreparedUpdate,
+  prepareUpdate,
 } from '../../../../repo'
-import { ids } from '../../../../lexicon/lexicons'
-import { Record as ProfileRecord } from '../../../../lexicon/types/app/bsky/actor/profile'
-import { ActorStoreTransactor } from '../../../../actor-store'
 
 export default function (server: Server, ctx: AppContext) {
   server.com.atproto.repo.putRecord({
