+++ packages/pds/src/context.ts
@@ -41,7 +41,7 @@ import { DidSqliteCache } from './did-cache'
 import { Crawlers } from './crawlers'
 import { DiskBlobStore } from './disk-blobstore'
 import { getRedisClient } from './redis'
-import { ActorStore } from './actor-store'
+import { ActorStore } from './actor-store/actor-store'
 import { LocalViewer, LocalViewerCreator } from './read-after-write/viewer'
 import { BskyAppView } from './bsky-app-view'
 import { ImageUrlBuilder } from './image/image-url-builder'
